add locations window
