+++ .gitignore
@@ -1,3 +1,4 @@
 .flatpak-builder
 .vscode
 build
+__pycache__
